Move public keys function to Role
