+++ conjur/api.py
@@ -220,65 +220,3 @@ class API(object):
         `api.resource('variable', 'db_password')`.
         """
         return Resource(self, kind, identifier)
-
-    def _public_key_url(self, *args):
-        return '/'.join([self.config.pubkeys_url] +
-                        [urlescape(arg) for arg in args])
-
-    def add_public_key(self, username, key):
-        """
-        Upload an openssh formatted public key to be made available for the user
-        given by `username`.
-
-        The key should be formatted like `ssh-rsa <data...> bob@example.com`.
-        """
-        self.post(self._public_key_url(username), data=key)
-
-    def remove_public_key(self, username, keyname):
-        """
-        Remove a specific public key for the user identified by `username`.
-        The `keyname` argument refers to the name field in the openssh formatted key
-        to be deleted.
-
-        For example, if they key contents are `ssh-rsa <data...> bob@example.com`,
-        the `keyname` should be `bob@example.com`
-        """
-        self.delete(self._public_key_url(username, keyname))
-
-    def remove_public_keys(self, username):
-        """
-        Remove all public keys for the user represented by `username`.
-        """
-        for keyname in self.public_key_names(username):
-            self.remove_public_key(username, keyname)
-
-    def public_keys(self, username):
-        """
-        Returns all keys for the user given by `username`, as a newline delimited string.
-
-        The odd format is chosen to support the Conjur SSH login implementation.
-        """
-        return self.get(self._public_key_url(username)).text
-
-    def public_key(self, username, keyname):
-        """
-        Return the contents of a specific public key given by `keyname`,
-        for the user given by `username` as a string.
-
-        The name of the key is based on the name entry of the openssh formatted key that was uploaded.
-
-        For example, if they key contents are `ssh-rsa <data...> bob@example.com`,
-        the `keyname` should be `bob@example.com`
-        """
-        return self.get(self._public_key_url(username, keyname)).text
-
-    def public_key_names(self, username):
-        """
-        Return the names of public keys for the user given by `username`.
-
-        The names of the keys are based on the name entry of the openssh formatted key that was uploaded.
-
-        For example, if they key contents are `ssh-rsa <data...> bob@example.com`,
-        the `keyname` should be `bob@example.com`
-        """
-        return [k.split(' ')[-1] for k in self.public_keys(username).split('\n')]
